Pdf_Extractor/requirements.txt
flask==2.0.1
PyMuPDF==1.18.14
pdf2image==1.16.0
pytesseract==0.3.8
pdfplumber==0.7.0
pandas==1.3.3
Pillow==8.3.2
Werkzeug==2.0.1 